requirements.txt
# Flask Framework
Flask==2.3.3
Flask-Session==0.5.0

# Database
pymongo==4.5.0

# Security
bcrypt==4.0.1
Werkzeug==2.3.7

# Machine Learning Dependencies (using pre-compiled wheels)
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.20.0
joblib>=1.1.0

# AI Chatbot
google-generativeai>=0.3.0

# Task Scheduler
APScheduler>=3.10.0

# Utilities
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
